// frontend/src/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
// ...
      events: {
        Row: {
          address: string | null
          category: string
          coordinates: unknown
          cover_image_url: string | null
          created_at: string
          currency: string
          description: Json
          end_date: string | null
          id: string
          is_free: boolean | null
          is_published: boolean
          location_name: string
          organizer_id: string | null
          original_lang: string
          price: number
          search_document: unknown
          start_date: string
          ticket_url: string | null
          title: Json
          updated_at: string
        }
        Insert: {
          address?: string | null
          category: string
          coordinates: unknown
          cover_image_url?: string | null
          created_at?: string
          currency?: string
          description: Json
          end_date?: string | null
          id?: string
          is_free?: boolean | null
          is_published?: boolean
          location_name: string
          organizer_id?: string | null
          original_lang: string
          price?: number
          search_document?: unknown
          start_date: string
          ticket_url?: string | null
          title: Json
          updated_at?: string
        }
        Update: {
          address?: string | null
          category?: string
          coordinates?: unknown
          cover_image_url?: string | null
          created_at?: string
          currency?: string
          description?: Json
          end_date?: string | null
          id?: string
          is_free?: boolean | null
          is_published?: boolean
          location_name?: string
          organizer_id?: string | null
          original_lang?: string
          price?: number
          search_document?: unknown
          start_date?: string
          ticket_url?: string | null
          title?: Json
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}